+++ packages/sapient-button/src/types.ts
@@ -1,0 +1,7 @@
+import { Variant } from '@sapiently/core';
+
+export type ButtonVariants = Exclude<Variant, 'warning'>;
+
+export type IconPosition = 'left' | 'right';
+export type ButtonShape = 'default' | 'circle';
+export type ButtonKind = 'default' | 'icon-only' | 'with-icon';