+++ packages/sapient-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sapiently/core
 
+## 0.3.0
+
+### Minor Changes
+
+- updated tokens, core, theme and button packages, button variants implemented and custom theming removed
+
 ## 0.2.1
 
 ### Patch Changes